Move RedwoodContent to Compose UI-specific module
While it doesn't have a Compose UI dependency now, it will gain one soon as we pipe HostConfiguration into it and need to pull things like dark mode, insets, etc.
