+++ redwood-compose/build.gradle
@@ -18,7 +18,6 @@ kotlin {
         api libs.kotlinx.coroutines.core
         api projects.redwoodRuntime
         api projects.redwoodWidget
-        implementation projects.redwoodWidgetCompose
         api libs.jetbrains.compose.runtime
       }
     }
