+++ redwood-composeui/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-composeui
+POM_NAME=Redwood for Compose UI
