+++ samples/counter/android-composeui/build.gradle
@@ -5,6 +5,7 @@ apply plugin: 'app.cash.redwood'
 dependencies {
   implementation projects.samples.counter.presenter
   implementation projects.samples.counter.sharedComposeui
+  implementation projects.redwoodComposeui
   implementation projects.redwoodLayoutComposeui
   implementation libs.kotlinx.coroutines.android
   implementation libs.google.material
