+++ samples/counter/desktop-composeui/build.gradle
@@ -13,5 +13,6 @@ dependencies {
   implementation compose.desktop.currentOs
   implementation projects.samples.counter.presenter
   implementation projects.samples.counter.sharedComposeui
+  implementation projects.redwoodComposeui
   implementation projects.redwoodLayoutComposeui
 }
