+++ settings.gradle
@@ -40,6 +40,7 @@ enableFeaturePreview('TYPESAFE_PROJECT_ACCESSORS')
 rootProject.name = 'redwood'
 
 include ':redwood-compose'
+include ':redwood-composeui'
 include ':redwood-flexbox'
 include ':redwood-gradle-plugin'
 include ':redwood-layout-api'
